[Lock][FrameworkBundle] Add LockBundle
(cherry picked from commit eece6cdbb95ecb6d05b246943a5f2b97ee759d28)
