+++ UPGRADE-8.2.md
@@ -94,6 +94,9 @@ FrameworkBundle
  * Deprecate the `framework.ide` config option, use the `SYMFONY_IDE` env var instead
  * BrowserKit assertions are no longer verbose by default. Failed response assertions no longer include the response body unless `setBrowserKitAssertionsAsVerbose(true)` is called or `verbose: true` is passed to the assertion.
  * Deprecate the `framework.fragments.hinclude_default_template` config option and the `fragment.renderer.hinclude.global_template` parameter; use the `esi` or `inline` fragment renderer, or [Symfony UX Turbo](https://ux.symfony.com/turbo), instead
+ * `LockBundle` is registered automatically when the Lock component is installed, and its services are then
+   registered without any configuration; set `lock.enabled` to `false` to disable them. The `framework.lock`
+   configuration is an alias of the `lock` configuration provided by the bundle
  * `SemaphoreBundle` is registered automatically when the Semaphore component is installed, and its services are then
    registered without any configuration; set `semaphore.enabled` to `false` to disable them. The `framework.semaphore`
    configuration is an alias of the `semaphore` configuration provided by the bundle
