+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/legacy_lock.php
@@ -1,5 +1,5 @@
 <?php
 
 $container->loadFromExtension('framework', [
-    'lock' => null,
+    'lock' => 'flock',
 ]);
