--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/lock_service_and_env.php
@@ -1,10 +1,0 @@
-<?php
-
-$container->setParameter('env(REDIS_DSN)', 'redis://paas.com');
-
-$container->loadFromExtension('framework', [
-    'lock' => [
-        'foo' => '%env(REDIS_DSN)%',
-        'bar' => 'my_service',
-    ],
-]);
