+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/legacy_lock.yml
@@ -1,0 +1,2 @@
+framework:
+    lock: flock
