--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/lock_advisory.yml
@@ -1,15 +1,0 @@
-services:
-    my_connection:
-        class: Doctrine\DBAL\Connection
-
-framework:
-    lock:
-        foo:
-            service_id: my_connection
-            advisory: true
-        bar:
-            service_id: my_connection
-        baz:
-            - flock
-            - service_id: my_connection
-              advisory: true
