--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/lock_named.yml
@@ -1,12 +1,0 @@
-parameters:
-    env(REDIS_DSN): redis://paas.com
-
-framework:
-    lock:
-        foo: semaphore
-        bar: flock
-        baz: [semaphore, flock]
-        qux: "%env(REDIS_DSN)%"
-        corge: in-memory
-        grault: mysql:host=localhost;dbname=test
-        garply: 'null'
