--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/lock_service.yml
@@ -1,6 +1,0 @@
-services:
-    my_service:
-        class: \Redis
-
-framework:
-    lock: my_service
