--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/lock_service_and_env.yml
@@ -1,8 +1,0 @@
-services:
-    my_service:
-        class: \Redis
-
-framework:
-    lock:
-      foo: "%env(REDIS_DSN)%"
-      bar: my_service
