+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTestCase.php
@@ -19,6 +19,7 @@ use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\Attributes\TestWith;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LogLevel;
+use Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\DefaultLockFactoryPass;
 use Symfony\Bundle\FrameworkBundle\DependencyInjection\FrameworkExtension;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\DummyMessage;
@@ -43,7 +44,6 @@ use Symfony\Component\DependencyInjection\Argument\TaggedIteratorArgument;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\Compiler\AddBehaviorDescribingTagsPass;
 use Symfony\Component\DependencyInjection\Compiler\ResolveBindingsPass;
-use Symfony\Component\DependencyInjection\Compiler\ResolveChildDefinitionsPass;
 use Symfony\Component\DependencyInjection\Compiler\ResolveTaggedIteratorArgumentPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -75,8 +75,7 @@ use Symfony\Component\HttpKernel\Exception\ConflictHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\ServiceUnavailableHttpException;
 use Symfony\Component\HttpKernel\Fragment\FragmentUriGeneratorInterface;
-use Symfony\Component\Lock\Store\FlockStore;
-use Symfony\Component\Lock\Store\SemaphoreStore;
+use Symfony\Component\Lock\LockBundle;
 use Symfony\Component\Mailer\EventListener\InMemoryPgpPublicKeyRepository;
 use Symfony\Component\Mailer\EventListener\InMemorySmimeCertificateRepository;
 use Symfony\Component\Mailer\EventListener\PgpMimeEncryptedMessageListener;
@@ -512,6 +511,15 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $this->assertSame('73902feb-9b95-4fe5-9c6f-b3e6d29e77b5', $definition->getArgument(5));
 
         $this->assertSame(Uuid47Transformer::class, $container->getDefinition('test_uuid47_transformer')->getClass());
+    }
+
+    public function testLockConfigurationIsForwardedToLockBundle()
+    {
+        $container = $this->createContainerFromFile('legacy_lock');
+
+        $this->assertTrue($container->hasDefinition('lock.default.factory'));
+        $this->assertSame('.lock.flock.store', (string) $container->getDefinition('lock.default.factory')->getArgument(0));
+        $this->assertSame('lock.default.factory', (string) $container->getAlias('lock.factory'));
     }
 
     public function testEnabledPhpErrorsConfig()
@@ -3385,114 +3393,6 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $this->assertSame('my-polyfill', $container->getDefinition('asset_mapper.importmap.renderer')->getArgument(3));
     }
 
-    public function testDefaultLock()
-    {
-        $container = $this->createContainerFromFile('lock');
-
-        $this->assertTrue($container->hasDefinition('lock.default.factory'));
-        $storeId = (string) $container->getDefinition('lock.default.factory')->getArgument(0);
-        $storeDef = $container->getDefinition($storeId);
-
-        if (class_exists(SemaphoreStore::class) && SemaphoreStore::isSupported()) {
-            $this->assertSame('.lock.semaphore.store', $storeId);
-            $this->assertSame(SemaphoreStore::class, $storeDef->getClass());
-            $this->assertSame('%kernel.project_dir%', $storeDef->getArgument(0));
-            $this->assertTrue($storeDef->hasTag('lock.store'));
-            $this->assertFalse($container->getDefinition('.lock.flock.store')->hasTag('lock.store'));
-        } else {
-            $this->assertSame('.lock.flock.store', $storeId);
-            $this->assertSame(FlockStore::class, $storeDef->getClass());
-            $this->assertTrue($storeDef->hasTag('lock.store'));
-            $this->assertFalse($container->getDefinition('.lock.semaphore.store')->hasTag('lock.store'));
-        }
-    }
-
-    public function testNamedLocks()
-    {
-        $container = $this->createContainerFromFile('lock_named');
-
-        $this->assertTrue($container->hasDefinition('lock.foo.factory'));
-        $storeId = (string) $container->getDefinition('lock.foo.factory')->getArgument(0);
-        $storeDef = $container->getDefinition($storeId);
-        $this->assertSame('.lock.semaphore.store', $storeId);
-        $this->assertSame(SemaphoreStore::class, $storeDef->getClass());
-        $this->assertSame('%kernel.project_dir%', $storeDef->getArgument(0));
-        $this->assertTrue($storeDef->hasTag('lock.store'));
-
-        $this->assertTrue($container->hasDefinition('lock.bar.factory'));
-        $storeId = (string) $container->getDefinition('lock.bar.factory')->getArgument(0);
-        $storeDef = $container->getDefinition($storeId);
-        $this->assertSame('.lock.flock.store', $storeId);
-        $this->assertSame(FlockStore::class, $storeDef->getClass());
-        $this->assertTrue($storeDef->hasTag('lock.store'));
-
-        $this->assertTrue($container->hasDefinition('lock.baz.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.baz.factory')->getArgument(0));
-        $this->assertIsArray($storeDefArg = $storeDef->getArgument(0));
-        $this->assertSame(['.lock.semaphore.store', '.lock.flock.store'], array_map('strval', $storeDefArg));
-
-        $this->assertTrue($container->hasDefinition('lock.qux.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.qux.factory')->getArgument(0));
-        $this->assertStringContainsString('REDIS_DSN', $storeDef->getArgument(0));
-
-        $this->assertTrue($container->hasDefinition('lock.corge.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.corge.factory')->getArgument(0));
-        $this->assertSame('in-memory', $storeDef->getArgument(0));
-
-        $this->assertTrue($container->hasDefinition('lock.grault.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.grault.factory')->getArgument(0));
-        $this->assertSame('mysql:host=localhost;dbname=test', $storeDef->getArgument(0));
-
-        $this->assertTrue($container->hasDefinition('lock.garply.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.garply.factory')->getArgument(0));
-        $this->assertSame('null', $storeDef->getArgument(0));
-    }
-
-    public function testLockWithService()
-    {
-        $container = $this->createContainerFromFile('lock_service', [], true, false);
-        $container->getCompilerPassConfig()->setOptimizationPasses([new ResolveChildDefinitionsPass()]);
-        $container->compile();
-
-        $this->assertTrue($container->hasDefinition('lock.default.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.default.factory')->getArgument(0));
-        $this->assertEquals(new Reference('my_service'), $storeDef->getArgument(0));
-    }
-
-    public function testLockWithAdvisoryService()
-    {
-        $container = $this->createContainerFromFile('lock_advisory', [], true, false);
-        $container->getCompilerPassConfig()->setOptimizationPasses([new ResolveChildDefinitionsPass()]);
-        $container->compile();
-
-        $storeDef = $container->getDefinition($container->getDefinition('lock.foo.factory')->getArgument(0));
-        $this->assertEquals([new Reference('my_connection'), true], $storeDef->getArguments());
-
-        $storeDef = $container->getDefinition($container->getDefinition('lock.bar.factory')->getArgument(0));
-        $this->assertEquals([new Reference('my_connection')], $storeDef->getArguments());
-
-        $combinedDef = $container->getDefinition($container->getDefinition('lock.baz.factory')->getArgument(0));
-        $this->assertIsArray($storeRefs = $combinedDef->getArgument(0));
-        $this->assertCount(2, $storeRefs);
-        $this->assertSame('.lock.flock.store', (string) $storeRefs[0]);
-        $this->assertEquals([new Reference('my_connection'), true], $container->getDefinition((string) $storeRefs[1])->getArguments());
-    }
-
-    public function testLockWithServiceAndEnv()
-    {
-        $container = $this->createContainerFromFile('lock_service_and_env', [], true, false);
-        $container->getCompilerPassConfig()->setOptimizationPasses([new ResolveChildDefinitionsPass()]);
-        $container->compile();
-
-        $this->assertTrue($container->hasDefinition('lock.foo.factory'));
-        $this->assertTrue($container->hasDefinition('lock.bar.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('lock.bar.factory')->getArgument(0));
-
-        $connection = $storeDef->getArgument(0);
-        $this->assertInstanceOf(Reference::class, $connection);
-        $this->assertEquals('my_service', $connection->__toString());
-    }
-
     public function testJsonStreamerEnabled()
     {
         $container = $this->createContainerFromFile('json_streamer');
@@ -3575,6 +3475,7 @@ abstract class FrameworkExtensionTestCase extends TestCase
         ], $data)));
 
         new ServicesBundle()->getContainerExtension()->load([], $container);
+        $container->registerExtension(new LockBundle()->getContainerExtension());
 
         return $container;
     }
@@ -3594,7 +3495,7 @@ abstract class FrameworkExtensionTestCase extends TestCase
             $container->getCompilerPassConfig()->setRemovingPasses([]);
             $container->getCompilerPassConfig()->setAfterRemovingPasses([]);
         }
-        $container->getCompilerPassConfig()->setBeforeOptimizationPasses([new AddBehaviorDescribingTagsPass(), new LoggerPass()]);
+        $container->getCompilerPassConfig()->setBeforeOptimizationPasses([new AddBehaviorDescribingTagsPass(), new LoggerPass(), new DefaultLockFactoryPass()]);
         $container->getCompilerPassConfig()->setBeforeRemovingPasses([new AddConstraintValidatorsPass(), new TranslatorPass()]);
 
         if (!$compile) {
@@ -3612,6 +3513,7 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $loader = new ClosureLoader($container);
         $loader->load($closure);
 
+        $container->addCompilerPass(new DefaultLockFactoryPass());
         $container->getCompilerPassConfig()->setOptimizationPasses([]);
         $container->getCompilerPassConfig()->setRemovingPasses([]);
         $container->getCompilerPassConfig()->setAfterRemovingPasses([]);
