+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -55,7 +55,7 @@
         "symfony/html-sanitizer": "^8.2",
         "symfony/http-client": "^8.2",
         "symfony/json-streamer": "^7.4|^8.0",
-        "symfony/lock": "^7.4|^8.0",
+        "symfony/lock": "^8.2",
         "symfony/mailer": "^7.4|^8.0",
         "symfony/messenger": "^7.4.10|^8.0.10",
         "symfony/mime": "^8.2",
@@ -95,6 +95,7 @@
         "symfony/json-path": "<8.2",
         "symfony/json-streamer": "<7.4",
         "symfony/mailer": "<7.4.17|>=8.0,<8.1.5",
+        "symfony/lock": "<8.2",
         "symfony/security-csrf": "<7.4",
         "symfony/semaphore": "<8.2",
         "symfony/serializer": "<7.4",
