+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `LockBundle`, which provides the `lock` configuration and the services previously provided by `FrameworkBundle` under `framework.lock`
  * Add argument `$advisory` to `StoreFactory::createStore()` to use advisory locks when reusing an existing `\PDO` or Doctrine DBAL `Connection`
  * Add `DoctrineDbalMysqlStore` based on MySQL `GET_LOCK()` functionality, usable with the `mysql+advisory://` DSN
  * Add `MysqlStore` based on MySQL `GET_LOCK()` functionality, usable with the `mysql+advisory:` DSN
