+++ src/Symfony/Component/Lock/Resources/config/lock.php
