+++ src/Symfony/Component/Lock/Tests/DependencyInjection/Fixtures/lock_advisory.php
@@ -1,0 +1,17 @@
+<?php
+
+$container->register('my_connection', \Doctrine\DBAL\Connection::class);
+
+$container->loadFromExtension('lock', [
+    'foo' => [
+        'service_id' => 'my_connection',
+        'advisory' => true,
+    ],
+    'bar' => [
+        'service_id' => 'my_connection',
+    ],
+    'baz' => [
+        'flock',
+        ['service_id' => 'my_connection', 'advisory' => true],
+    ],
+]);
