+++ src/Symfony/Component/Lock/Tests/DependencyInjection/Fixtures/lock_disabled.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('lock', ['enabled' => false]);
