+++ src/Symfony/Component/Lock/Tests/DependencyInjection/Fixtures/lock_named.php
@@ -1,0 +1,13 @@
+<?php
+
+$container->setParameter('env(REDIS_DSN)', 'redis://paas.com');
+
+$container->loadFromExtension('lock', [
+    'foo' => 'semaphore',
+    'bar' => 'flock',
+    'baz' => ['semaphore', 'flock'],
+    'qux' => '%env(REDIS_DSN)%',
+    'corge' => 'in-memory',
+    'grault' => 'mysql:host=localhost;dbname=test',
+    'garply' => 'null',
+]);
