+++ src/Symfony/Component/Lock/Tests/DependencyInjection/Fixtures/lock_service.php
@@ -1,0 +1,5 @@
+<?php
+
+$container->register('my_service', \Redis::class);
+
+$container->loadFromExtension('lock', ['default' => 'my_service']);
