+++ src/Symfony/Component/Lock/Tests/DependencyInjection/Fixtures/lock_service_and_env.php
@@ -1,0 +1,8 @@
+<?php
+
+$container->setParameter('env(REDIS_DSN)', 'redis://paas.com');
+
+$container->loadFromExtension('lock', [
+    'foo' => '%env(REDIS_DSN)%',
+    'bar' => 'my_service',
+]);
