+++ src/Symfony/Component/Lock/composer.json
@@ -22,10 +22,13 @@
     "require-dev": {
         "doctrine/dbal": "^4.3",
         "predis/predis": "^1.1|^2.0",
+        "symfony/config": "^7.4|^8.0",
+        "symfony/dependency-injection": "^8.1",
         "symfony/serializer": "^6.4|^7.0"
     },
     "conflict": {
-        "doctrine/dbal": "<4.3"
+        "doctrine/dbal": "<4.3",
+        "symfony/dependency-injection": "<8.1"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Lock\\": "" },
